fix: types and properties names
