+++ packages/ocean-react/src/CrossSellCard/CrossSellCard.tsx
@@ -1,33 +1,63 @@
-import React, { forwardRef } from 'react';
+import React, { forwardRef, ComponentPropsWithoutRef } from 'react';
 import { ChevronRight } from '@useblu/ocean-icons-react';
 import classNames from 'classnames';
 
-import { CrossSellCardProps } from './types';
+export type CrossSellCardProps = {
+  /**
+   * Determines the card title
+   */
+  title: string;
+  /**
+   * Determines the title color
+   */
+  titleColor?: string;
+  /**
+   * Determines the card description
+   */
+  description: string;
+  /**
+   * Determines the description color
+   */
+  descriptionColor?: string;
+  /**
+   * Determines the card image src
+   */
+  imageSrc?: string;
+  /**
+   * Determines the cta text
+   */
+  ctaText: string;
+  /**
+   * Determines the Icon on the
+   */
+  customIcon?: JSX.Element;
+  /**
+   * Determines the cta action
+   */
+  ctaAction: () => void;
+  /**
+   * Determines the background color
+   */
+  backgroundColor?: string;
+} & ComponentPropsWithoutRef<'div'>;
 
 const CrossSellCard = forwardRef<HTMLDivElement, CrossSellCardProps>(
   function CrossSellCard(
     {
       title,
-      className,
       description,
+      titleColor = 'white',
+      descriptionColor = '#b8c3ff',
+      className,
       imageSrc,
       ctaText,
       ctaAction,
       customIcon,
-      backgroundColor,
+      backgroundColor = '#0025e0',
       ...rest
     },
     ref
   ) {
-    const titleCard: string = typeof title === 'string' ? title : title['text'];
-    const descriptionCard: string =
-      typeof description === 'string' ? description : description['text'];
-
-    const titleColor: string =
-      typeof title === 'string' ? 'white' : title['color'];
-    const descriptionColor: string =
-      typeof description === 'string' ? '#b8c3ff' : description['color'];
-
     return (
       <div
         ref={ref}
@@ -39,20 +69,20 @@ const CrossSellCard = forwardRef<HTMLDivElement, CrossSellCardProps>(
           className="ods-cross-sell-card__content"
           onClick={() => ctaAction}
           data-testid="cta-test"
-          style={{ background: backgroundColor || '#0025e0' }}
+          style={{ background: backgroundColor }}
         >
           <div className="ods-cross-sell-card__information">
             <div
               className="ods-typography--inverse ods-typography__heading3"
               style={{ color: titleColor }}
             >
-              {titleCard}
+              {title}
             </div>
             <div
               className="ods-typography--inverse ods-typography__description ods-cross-sell-card__description"
               style={{ color: descriptionColor }}
             >
-              {descriptionCard}
+              {description}
             </div>
           </div>
           {imageSrc && (
